fix pmm

+++ arch/i386/kernel/vmm/vmm.c
@@ -4,22 +4,22 @@
 extern uint32_t pmm_alloc_page(void);
 extern void pmm_free_page(uint32_t physical_addr);
 
-static uint32_t *page_directory = 0;
+static uintptr_t *page_directory = 0;
 
 void vmm_init() {
-    page_directory = (uint32_t *) pmm_alloc_page();
+    page_directory = (uintptr_t *) pmm_alloc_page();
 
     for (int i = 0; i < 1024; i++) {
         page_directory[i] = 0x00000002;
     }
 
-    uint32_t *first_page_table = (uint32_t *) pmm_alloc_page();
+    uintptr_t *first_page_table = (uintptr_t *) pmm_alloc_page();
 
     for (int i = 0; i < 1024; i++) {
         first_page_table[i] = (i * 4096) | 3; 
     }
 
-    page_directory[0] = ((uint32_t) first_page_table) | 3;
+    page_directory[0] = ((uintptr_t) first_page_table) | 3;
 
     __asm__ volatile (
         "mov %0, %%cr3\n\t"
